+++ src/cmd7.c
@@ -1279,7 +1279,7 @@ int show_flags(byte group, int pval)
 				sprintf(ttt, "%c) %s (exp %ld)",
 				        (items < 26) ? I2A(items) : ('0' + items - 26),
 				        al_name + a_select_flags[i].desc,
-				        a_select_flags[i].xp);
+				        (long int) a_select_flags[i].xp);
 
 			/* Note: Somebody is VERY clever, and it wasn't me. Text printed as
 			 * TERM_DARK is actually printed as TERM_BLUE *SPACES* to prevent the
